backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
python-multipart
email-validator
crewai
crewai-tools
langchain-openai
